requirements.txt
fastapi
uvicorn
tensorflow
pillow
huggingface_hub
python-multipart
streamlit
